backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.14
Django==4.2.4
django-cors-headers==4.2.0
dnspython==2.4.1
docopt==0.6.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.253
langsmith==0.0.19
loguru==0.7.0
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.2
pipreqs==0.4.13
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
yarg==0.1.9
yarl==1.9.2
